fix: move system prompt from test file to production code
Addresses Greptile feedback - production code should not import from test files.

Changes:
- Created apps/web/src/lib/jon-mode-prompts.ts for EM_DASH_PREVENTION_SYSTEM_PROMPT
- Updated all imports across:
  - chat-handler.ts (production code)
  - jon-mode-prompts.test.ts (test file)
  - jon-mode-integration.test.ts
  - jon-mode-simple.test.ts
- Maintains architectural separation between production and test code
- All tests pass, build successful
